[BACKLOG-2100] Remove Hadoop Clusters from Kettle
